backend/requirements-railway.txt
# Minimal requirements for Railway deployment
Flask==3.1.2
Flask-CORS==6.0.1
python-dotenv==1.1.1
Pillow==11.3.0
gunicorn==21.2.0